Doc improvements for custom processors (#410)
* Added the module load message to return error

I have passed the module loading error message on to the user to
help aid them in debugging their issues.  The underlying error
will now be returned via the API to the submitter.

* added custom_operations, renamed reference updated README

* Handling perserve_id for generated data

elasticsearch_index_selector now copies the id from the document
to the meta info when preserve_id is set.  id is then deleted
from the document.

TODO: still figure out how to handle the post_validation.

* elasticsearch_data_generator is exempt from preserve_id check

* fixed formatting

* added test for post_validation

* minor doc improvlemt

* undoing the changes I made the the index selector and tests