Extract validateOperation from validators.job
To support development of the Teraslice Test harness, specifically
https://github.com/terascope/teraslice_op_test_harness/issues/2

I have extracted validateOperation.  It is renamed to
validateConfig since it handles both jobConfigs and opConfigs.

I have also removed the third argument, which was a boolean
that indicated whether the config passes was an operation or not.
I replaced it with a test for the presence of ._op which should
be present on all operations.