Staggering worker recycles randomly
When the user sets recycle_worker on a job the workers are
recycled after they hit that slice count.  These can tend to
happen around the same time for all workers.  In the case of
kafka workers, the time it takes to rebalance partitions takes
some time.  By recycling so many workers in a short period of
time we end up accumulating a lot of lag.

This PR tries to stagger the recycle times by comparing
sliceCount with somewhere between 75% and 100% of the
random_worker setting.

Since this is called at the end of every slice, I don't think
it's going to work the way I had hoped.  I think it will result
in most workers recycling near the 75% mark.