Ensure execution_controller pods are remove on job stop
This is called the Cascading Deletion Policy by k8s.  Basically we just
need to pass a body with the request that will ensure pods get deleted
too.  I suppose we might encounter conditions where we don't want the
pods deleted when the job is, but for now, we'll force that.

closes: #777