Execution controller can exit when active workers are gone
I am not sure if this is actually the right logic, but maybe it's OK
This commit will allow the execution_controller to shutdown when
there are no more active workers connected.  I am pretty sure
we can only get into this loop during shutdown.  But there's no
other functional mechanism to exit this loop.