fix: change the default production/test index replicas / shards in the IndexModel
Previously this would default to 1 replica and 5 shards (if NODE_ENV is empty or NODE_ENV === production). Now the default is 2 replica and 5 shards (if NODE_ENV !== 'test'). This will prevent accidentally creating indexes with less replicas or shards if NODE_ENV is set to development, like if you are running a migration or running against a production instance.