Update error message to match tests
I have changed my error message to match those expected
to be raised in the tests.  I think this is consistent with
the expectations of the tests.  They want an error to be
raised when the URL contains an invalid hostname.  Technically
the error is raised from different code, but I think it
serves the same purpose.