Remove PromMetrics proxy, check for promMetrics with isPromAvailable() (#3627)
This is a follow-up PR to #3629 which created a utility function to test
if promMetrics are available, allowing assets to be backward compatible
with teraslice before promMetrics were introduced (< 1.4.0). At the same
time we decided to just make `isPromAvailable()` checks everywhere
promMetrics was used, instead of using a proxy to catch calls to
undefined functions. See #3625 for more details.

This PR makes the following changes:
- Removes the `promMetricsAPI` Proxy from `terafoundation`
- Add checks using `isPromAvailable` before all calls to
`promMetricsAPI` functions.
- Update `promMetrics docs`.
- Bump teraslice from 1.7.1 to 1.7.2
- Bump job-components from 0.75.0 to 0.75.1
- Bump terafoundation from 0.64.0 to 0.64.1