Set default label 'name' at init, not when PromMetrics class is created. (#3647)
The default label `name` was being set at `PromMetrics` class creation
within `terafoundation` instead of being set at initialization. It
previously was getting the value for `name` from `context.name`, but
this doesn't always match with `teraslice.name` and would never be
correct if used with `spaces`.

There is now a test in `prom-metrics-spec.ts` that checks the default
labels for the correct values.