revert @swc/core to version 1.4.0 (#3661)
When upgrading `@swc/core` from v1.4.0 to v1.5.5 there was a change that
results in source mappings failing. This would print thousands of lines
of error messages in out e2e tests, but everything seems to compile
okay. To cut down this noise until there is a fix, this PR reverts the
`@swc/core` version back to 1.4.0.

More details and links here: #3655