[terafoundation] Prom metrics exporter doesn't reset metrics between updates (#3747)
This PR makes the following changes:

- The `PromMetrics` class needs to reset it's list of metrics on each
scrape. If it doesn't do this, then all the executions are listed, not
just the active ones. `resetMetrics()` functions were added to
`PromMetrics` and `Exporter` to reset the `prom-client` register.
- Add `prom_metrics_display_url` field to terafoundation. This value
will be used as the `url` default label added to all prom metrics.
Defaults to an empty string, making it more obvious that this field is
missing from the config.
- Include cluster analytics metrics (GET '/cluster/stats' endpoint
results) in the cluster master

ref: #3743