[workspace] Set Jest transformIgnorePatterns to default (#3880)
This PR makes the following changes:
- In `jest.config.base.js` explicitly set `transformIgnorePatterns` to
the default (`['/node_modules/', '\\.pnp\\.[^\\/]+$']`), which will skip
running transforms on node-modules or pnp files. We were overriding the
default with an empty array. Standard practice is to only transform
specific dependencies if there is a need.
- Remove patch for `unicorn-magic` dependency, as it is no longer being
broken by the jest transform.
- Add README to `./patches` explaining how to make a patch using
`patch-package`. This also prevents the `patches` directory from being
removed if there are no patches, breaking the docker build.

Ref: #3871, #3874