[workspace] add workflow job to automate Teraslice releases (#3895)
This PR makes the following changes:
- Add a new job to the `publish-master.yml` workflow called
`create-release-on-teraslice-bump`. It first uses `semver-compare-cli`
to check if the version of Teraslice in the root `package.json` is
greater than it was in the previous master branch. If it has then the
`softprops/action-gh-release@v2` action creates a release. As with a
manual release, the publish-tag.yml workflow will automatically run when
the release is created.
- Add a test to ensure the teraslice version and root package.json
version stay in sync.

TODO: Should we bump teraslice just to make the new workflow job run?

ref: #1502