[e2e, helm] autoload_directory should be created in e2e helm, not be default in helm chart (#4006)
This PR fixes a mistake that added `autoload_directory` to the default
teraslice helm `values.yaml`. This setting is only needed for e2e, so is
now set in `e2e/helm/templates/teraslice.yaml.gotmpl`.

The chart is bumped to version 2.3.1

I also added `icon` and `home` links to the chart.