[CI] V3 release pipeline (#4181)
This PR makes the following changes:
- add `semver-tools.sh` script
- update release workflows
- use `semver-tool.sh` in `publish-master.yml` instead of
`semver-compare-cli`. This will fix merges to master with a new semver
pre-releases tag not being released #3946
- `publish-dev.yml` is triggered on a merge to `release/teraslice-v*`
and does the following:
- `ts-scripts publish -t prerelease npm` will publish all packages that
currently have an updated pre-release version.
- `ts-scripts publish -t prerelease docker` will publish node 22/24
docker images only if teraslice has a pre-release version and an image
with that tag does not already exist.
- create a github release if the teraslice version has increased and it
is a pre-release.
- create a slack announcement that a new pre-release has been published.
- `publish-tag.yml` will not publish to npm or docker if the teraslice
version is a prerelease
 

ref: #4176