[CI] Create dev release on right branch (#4195)
This PR makes the following changes:
- add `target_commitish: ${{ github.event.pull_request.base.ref }}` to
github release action.
- From the docs: "Commitish value that determines where the Git tag is
created from. Can be any branch or commit SHA. Defaults to repository
default branch."
https://github.com/softprops/action-gh-release/tree/v2/?tab=readme-ov-file#-customizing
- `github.event.pull_request.base.ref` will be the branch the PR was
just merged into, so this will be the new merge commit.
  - This will fix all current bugs:
- git tag on master branch (because that's the default the action used)
- release assets are master branch code (because that's the default the
action used)
- `publish-tag.yml` workflow using master branch version to check if
release was a prerelease (because that's the code the release had in it)
- add `ref: ${{ github.event.pull_request.base.ref }}` when checking out
code.
- This is just a precaution, as the correct branch was checked out by
default, but making it explicit never hurts
- bump version to `dev.4`