[teraslice, teraslice-messaging] Socket io upgrade (#4201)
This PR makes the following changes:
- bump socket.io and socket.io-client from v1.7.4 to v4.8.1
- bump teraslice from v3.0.0-dev.5 to v3.0.0-dev.6
- bump teraslice-messaging from v2.0.0-dev.5 to v2.0.0-dev.6
- remove @types/socket.io and @types/socket.io-client as types are now
built in
- the `socket` client class no longer forwards emitted events from the
`socket.io` class (Manager) so many events are now listened for with
`socket.io.on()` instead of `socket.on()`
- The manager now emits generic `error` events as well as specific
`connection_*` events. These are logged at trace to avoid crowding the
logs.
- `connect_timeout` and `reconnecting` events are no longer emitted. We
listen for `connect_error` and `reconnect_attempt` instead
- The socket.io server `maxHttpBufferSize` now defaults to `1MB`. It is
now explicitly set to the v1 default of `100MB`
- `socket.handshake.query` is now `socket.handshake.auth`

refs:
- #3467
- https://socket.io/docs/v4/migrating-from-2-x-to-3-0/
- https://socket.io/docs/v4/migrating-from-3-x-to-4-0/